fix: fix racy access to the provider ID variable
Because the provider ID was set on a package level and was accessed by package-level variables from different packages, there were cases where it was accessed before it was written. This caused the provider to not respect the custom IDs passed via `--id` flag.

Address this by delaying access to the ID by replacing some package level vars with functions, e.g., the resource types.

Also add a safeguard to panic when the provider ID is accessed, but it is not initialized yet.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>