fix: fix pxe boot mode change logic
Our logic for rebooting to change modes was wrong: after responding to an iPXE request, we immediately marked the machine as booted into the desired mode.

In fact, there is no guarantee that that machine will actually be able to pull the kernel/initramfs etc and boot into the mode decided by the iPXE handler. Things can go wrong (e.g., image factory connectivity problems, provider restarts etc.).

In these cases, we wrongly marked the machine as "booted into mode X", and since we relied on that information to decide whether to reboot the machine or not later, we sometimes did not issue required reboots.

Fix this my simplifying the logic:
- If the machine needs to be in the agent mode, but it isn't, issue a reboot
- If the machine needs to be not in the agent mode, but it is, issue a reboot

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>