feat: allow providing own ca cert and self-issued certs
Add a new mode to be able to give the provider a CA cert, a cert and a key to be used in TLS mode.

This allows users to be able to use a single CA for Omni and the provider, and/or use a public CA to sign their certificates.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>